+++ postgrest.cabal
@@ -47,6 +47,7 @@ library
                       PostgREST.App
                       PostgREST.AppState
                       PostgREST.AppState.Pool
+                      PostgREST.AppState.Reload
                       PostgREST.AppState.Types
                       PostgREST.Auth
                       PostgREST.Auth.Jwt
@@ -70,7 +71,6 @@ library
                       PostgREST.SchemaCache.Table
                       PostgREST.Error
                       PostgREST.Error.Types
-                      PostgREST.Listener
                       PostgREST.Logger
                       PostgREST.MainTx
                       PostgREST.Logger.Apache
